Change to Double

+++ Database/Updates/World/2019-01-20-00-Spell_Table_DotDuration_Add.sql
@@ -1,4 +1,4 @@
 USE `ace_world`;
 
 ALTER TABLE `spell` 
-ADD COLUMN `dot_Duration` INT(10) NULL DEFAULT NULL AFTER `number_Variance`;
+ADD COLUMN `dot_Duration` DOUBLE NULL DEFAULT NULL AFTER `number_Variance`;
